+++ CHANGELOG.md
@@ -14,6 +14,11 @@ For older changes see the [archived Singularity change log](https://github.com/a
 - When building RPM, we will now use `/var/lib/apptainer` (rather than
   `/var/apptainer`) to store local state files.
 
+### New Features & Functionality
+
+- The `remote status` command will now print the username, realname, and email
+  of the logged-in user, if available.
+
 ## Changes since last pre-release
 
 - Upgrade gocryptfs to version 2.4.0, removing the need for fusermount from